Update commandtool.bat
Added ability to add network locations to the tool for quick access. Also added an experimental feature to write notes for you based on the commands selected. Notes will automatically open when you try to change the target, and will automatically delete themselves to save space, so make sure to copy them before closing the window!