Update commandtool.bat
General cleanup of the script. Added the ability to add an argument if launching from command line to select local mode or computer name. Was meant to be used as a way to preserve targets across restarts, but that is no longer necessary and now only exists for testing.

Added local mode for local troubleshooting. While this is enabled, psexec commands will instead be run normally on the local machine, and psexec/pskill exclusive commands will not be available. Added a toggle to the main menu for this. Local mode will preserve the target you have selected when you enable and disable it, so you can safely switch back and forth between the modes.